pyproject.toml
[project]
name = "mario-kart"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mss>=10.0.0",
    "numpy>=2.3.2",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "pygetwindow>=0.0.9",
    "pynput>=1.8.1",
    "stable-baselines3>=2.7.0",
]
